requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
python-dotenv
twilio
psycopg2-binary
python-multipart

mem0ai